achieved some code reuse
