+++ test/visual/resize_filter.js
@@ -1,84 +1,23 @@
 (function() {
   fabric.enableGLFiltering = false;
   fabric.isWebglSupported = false;
-  var _pixelMatch;
-  var visualCallback;
-  var fs;
+  fabric.Object.prototype.objectCaching = false;
+  var visualTestLoop;
+  var getFixture;
   if (fabric.isLikelyNode) {
-    fs = global.fs;
-    _pixelMatch = global.pixelmatch;
-    visualCallback = global.visualCallback;
+    visualTestLoop = global.visualTestLoop;
+    getFixture = global.getFixture;
   }
   else {
-    _pixelMatch = pixelmatch;
-    if (window) {
-      visualCallback = window.visualCallback;
-    }
+    visualTestLoop = window.visualTestLoop;
+    getFixture = window.getFixture;
   }
   var fabricCanvas = this.canvas = new fabric.Canvas(null, {enableRetinaScaling: false, renderOnAddRemove: false});
-  var pixelmatchOptions = {
-    includeAA: false,
-    threshold: 0.095
-  };
-  fabric.Object.prototype.objectCaching = false;
-
-  function getAbsolutePath(path) {
-    var isAbsolute = /^https?:/.test(path);
-    if (isAbsolute) { return path; };
-    var imgEl = fabric.document.createElement('img');
-    imgEl.src = path;
-    var src = imgEl.src;
-    imgEl = null;
-    return src;
-  }
-
-  function getFixtureName(filename) {
-    var finalName = '/fixtures/' + filename;
-    return fabric.isLikelyNode ? (__dirname + '/..' + finalName) : getAbsolutePath('/test' + finalName);
-  }
-
-  function getGoldeName(filename) {
-    var finalName = '/golden/' + filename;
-    return fabric.isLikelyNode ? (__dirname + finalName) : getAbsolutePath('/test/visual' + finalName);
-  }
-
-  function getImage(filename, original, callback) {
-    if (fabric.isLikelyNode && original) {
-      try {
-        fs.statSync(filename);
-      }
-      catch (err) {
-        var dataUrl = original.toDataURL().split(',')[1];
-        console.log('creating original for ', filename);
-        fs.writeFileSync(filename, dataUrl, { encoding: 'base64' });
-      }
-    }
-    var img = fabric.document.createElement('img');
-    img.onload = function() {
-      img.onload = null;
-      callback(img);
-    };
-    img.onerror = function(err) {
-      img.onerror = null;
-      console.log('Image loading errored', err);
-    };
-    img.src = filename;
-  }
-
-  function beforeEachHandler() {
-    fabricCanvas.setZoom(1);
-    fabricCanvas.setDimensions({
-      width: 300,
-      height: 150,
-    });
-    fabricCanvas.clear();
-    fabricCanvas.renderAll();
-  }
 
   var tests = [];
 
   function imageResizeTest(canvas, callback) {
-    getImage(getFixtureName('parrot.png'), false, function(img) {
+    getFixture('parrot.png', false, function(img) {
       canvas.setDimensions({
         width: 200,
         height: 200,
@@ -100,11 +39,11 @@
     code: imageResizeTest,
     golden: 'parrot.png',
     newModule: 'Image resize filter test',
-    percentage: 0.06,
+    percentage: 0.08,
   });
 
   function imageResizeTestNoZoom(canvas, callback) {
-    getImage(getFixtureName('parrot.png'), false, function(img) {
+    getFixture('parrot.png', false, function(img) {
       canvas.setDimensions({
         width: 200,
         height: 200,
@@ -123,11 +62,11 @@
     test: 'Image resize without zoom',
     code: imageResizeTestNoZoom,
     golden: 'parrot.png',
-    percentage: 0.06,
+    percentage: 0.08,
   });
 
   function imageResizeTestGroup(canvas, callback) {
-    getImage(getFixtureName('parrot.png'), false, function(img) {
+    getFixture('parrot.png', false, function(img) {
       canvas.setDimensions({
         width: 200,
         height: 200,
@@ -148,11 +87,11 @@
     test: 'Image resize with scaled group',
     code: imageResizeTestGroup,
     golden: 'parrot.png',
-    percentage: 0.06,
+    percentage: 0.08,
   });
 
   function blendImageTest2(canvas, callback) {
-    getImage(getFixtureName('parrot.png'), false, function(img) {
+    getFixture('parrot.png', false, function(img) {
       var image = new fabric.Image(img);
       var backdropImage = new fabric.Image(img);
       backdropImage.left = backdropImage.width;
@@ -181,8 +120,8 @@
   });
 
   function blendImageTest(canvas, callback) {
-    getImage(getFixtureName('parrot.png'), false, function(img) {
-      getImage(getFixtureName('very_large_image.jpg'), false, function(backdrop) {
+    getFixture('parrot.png', false, function(img) {
+      getFixture('very_large_image.jpg', false, function(backdrop) {
         var image = new fabric.Image(img);
         var backdropImage = new fabric.Image(backdrop);
         image.filters.push(new fabric.Image.filters.BlendImage({image: backdropImage, alpha: 0.5 }));
@@ -206,48 +145,5 @@
     percentage: 0.06,
   });
 
-  tests.forEach(function(testObj) {
-    var testName = testObj.test;
-    var code = testObj.code;
-    var percentage = testObj.percentage;
-    var golden = testObj.golden;
-    var newModule = testObj.newModule;
-    if (newModule) {
-      QUnit.module(newModule, {
-        beforeEach: beforeEachHandler,
-      });
-    }
-    QUnit.test(testName, function(assert) {
-      var done = assert.async();
-      code(fabricCanvas, function(renderedCanvas) {
-        var width = renderedCanvas.width;
-        var height = renderedCanvas.height;
-        var totalPixels = width * height;
-        var imageDataCanvas = renderedCanvas.getContext('2d').getImageData(0, 0, width, height).data;
-        var canvas = fabric.document.createElement('canvas');
-        canvas.width = width;
-        canvas.height = height;
-        var ctx = canvas.getContext('2d');
-        var output = ctx.getImageData(0, 0, width, height);
-        getImage(getGoldeName(golden), renderedCanvas, function(goldenImage) {
-          ctx.drawImage(goldenImage, 0, 0);
-          visualCallback.addArguments({
-            enabled: true,
-            golden: canvas,
-            fabric: renderedCanvas,
-            diff: output
-          });
-          var imageDataGolden = ctx.getImageData(0, 0, width, height).data;
-          var differentPixels = _pixelMatch(imageDataCanvas, imageDataGolden, output.data, width, height, pixelmatchOptions);
-          var percDiff = differentPixels / totalPixels * 100;
-          var okDiff = totalPixels * percentage;
-          assert.ok(
-            differentPixels < okDiff,
-            testName + ' has too many different pixels ' + differentPixels + '(' + okDiff + ') representing ' + percDiff + '%'
-          );
-          done();
-        });
-      });
-    });
-  });
+  tests.forEach(visualTestLoop(fabricCanvas, QUnit));
 })();
